+++ task/lib/src/widgets/task_list_view.dart
@@ -57,6 +57,7 @@ class TaskListView extends StatelessWidget {
                           '${(task.id == 0) ? '-' : task.id} '
                                   '${pendingFilter ? '' : '${task.status[0].toUpperCase()} '}'
                                   '${age(task.entry)} '
+                                  '${(task.start != null) ? 'st:${age(task.start!)}' : ''} '
                                   '${(task.due != null) ? 'd:${when(task.due!)}' : ''} '
                                   '${task.priority ?? ''} '
                                   '${task.project ?? ''} '